Renamed user command handler
