+++ src/main/java/espresso/EspressoConfig.java
@@ -1,7 +1,9 @@
 package espresso;
 
+import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.ResourceBundleMessageSource;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -25,5 +27,13 @@ public class EspressoConfig implements WebMvcConfigurer {
                         .title("Achievement Unlocked Project - Espresso - Core")
                         .description("These are the core services of the espresso project.")
                         .version("1.0.0"));
+    }
+
+    @Bean
+    MessageSource messageSource() {
+        ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
+        messageSource.setBasename("messages/achievement_api_messages");
+        messageSource.setDefaultEncoding("UTF-8");
+        return messageSource;
     }
 }
